Moving model files into separate folder
