+++ web_app/app/models.py
@@ -1,42 +1,9 @@
 from web_app.app import db
-
-class User(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    t_screen_name = db.Column(db.String(120), index=True, unique=True)
-    
-    contributors = db.Column(db.Boolean)
-    created_at = db.Column(db.DateTime(True))
-    description = db.Column(db.String(400))
-    favourites_count = db.Column(db.Integer)
-    followers_count = db.Column(db.Integer)
-    following = db.Column(db.Boolean)
-    friends_count = db.Column(db.Integer)
-    t_id = db.Column(db.Integer)
-    name = db.Column(db.String(30))
-    statuses_count = db.Column(db.Integer)
-    verified = db.Column(db.Boolean)
-
-    def __repr__(self):
-        return '<User %r>' % (self.nickname)
-    
-    @staticmethod
-    def make_unique_nickname(nickname):
-        if User.query.filter_by(nickname=nickname).first() is None:
-            return nickname
-        version = 2
-        while True:
-            new_nickname = nickname + str(version)
-            if User.query.filter_by(nickname=nickname).first() is None:
-                break
-            version += 1
-        return new_nickname
-    
-class Follows(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    follower = db.Column(db.Integer, db.ForeignKey('user.t_userame'), index=True)
-    followee = db.Column(db.Integer, db.ForeignKey('user.t_userame'), index=True)
+from web_app.app.models.user_model import User
+from web_app.app.models.twitter_model import Follows, Status, Hashtag, Tags
     
 class Post(db.Model):
+    # This is a post from the tutorial
     id = db.Column(db.Integer, primary_key = True)
     title = db.Column(db.String(80))
     body = db.Column(db.String(140))
@@ -45,16 +12,3 @@ class Post(db.Model):
 
     def __repr__(self):
         return '<Post %r>' % (self.body)
-    
-class Status(db.Model):
-    id = db.Column(db.Integer, primary_key = True)
-    
-    favorited = db.Column(db.Boolean)
-    created_at = db.Column(db.DateTime)
-    
-class Tags(db.Model):
-    id = db.Column(db.Integer, primary_key=True)
-    status = db.Column(db.Integer, db.ForeignKey(''))
-    
-class Hashtag(db.Model):
-    text = db.Column(db.String, primary_key = True)
